Update moduledoc for Battery module
The intention here is to make the moduledoc more scannable at a glance.
It's tough to read paragraphs during a pairing interview.

I tried using the same Markdown syntax they use in the moduledocs for
the Elixir source code (the link to the Enum source code is linked below
for reference). But I wonder if there's still some improvements we can
make here to make this easier to scan at a glance without being
overwhelming.

https://github.com/elixir-lang/elixir/blob/master/lib/elixir/lib/enum.ex